search: handle spaces
Prior this PR, searching for "function plot" in the search bar resulted in
"function%20plot" in the glossary page's search box.
This PR fixes the issue by using `URLSearchParams`, which automatically
calls `decodeURIComponent` for us. We also no longer need to parse the
query string to a map ourselves.

`URLSearchParams` is supported in most major browsers since 2016/2017.

Fixes #75